refactor: 설정 페이지 수정
