+++ src/app/pages/Home.js
@@ -2,10 +2,13 @@ import React, { useState, useEffect } from "react";
 import UserService from "../services/user.service";
 import Lecture from "../components/Lecture/Lecture";
 import LECTURE_DATA from "../data/lectureData";
+import { useSelector } from "react-redux";
 
 const Home = () => {
   const accessToken = localStorage.getItem("accessToken");
   const userName = localStorage.getItem("username");
+
+  const { isLoggedIn } = useSelector((state) => state.auth);
 
   const [data, setData] = useState([]);
   useEffect(() => {
@@ -13,7 +16,7 @@ const Home = () => {
   }, []);
 
   useEffect(() => {
-    if (accessToken) {
+    if (isLoggedIn) {
       UserService.getUserProfile(accessToken).then((response) => {
         const { username } = response.data;
         if (!userName) {
@@ -21,7 +24,7 @@ const Home = () => {
         }
       });
     }
-  }, [userName, accessToken]);
+  }, [userName, accessToken, isLoggedIn]);
 
   return (
     <div className="home-page">
