--- src/app/pages/Settings.js
@@ -1,164 +1,0 @@
-import React, { useState, useEffect } from "react";
-import UserService from "../services/user.service";
-import { useDispatch, useSelector } from "react-redux";
-import { Navigate } from "react-router-dom";
-import { Formik, Field, Form, ErrorMessage } from "formik";
-import * as Yup from "yup";
-
-import { editProfile } from "../slices/auth";
-import { clearMessage } from "../slices/message";
-
-const Settings = () => {
-  const accessToken = localStorage.getItem("accessToken");
-  const userName = localStorage.getItem("username");
-  const [user, setUser] = useState("");
-
-  const { email, username } = user;
-
-  useEffect(() => {
-    UserService.getUserProfile(accessToken).then(
-      (response) => {
-        setUser(response.data);
-        const { username } = response.data;
-        if (!userName) {
-          localStorage.setItem("username", username);
-        }
-      },
-      (error) => {
-        const _content =
-          (error.response && error.response.data) ||
-          error.message ||
-          error.toString();
-        setUser(_content);
-      }
-    );
-  }, [accessToken, userName]);
-
-  const [loading, setLoading] = useState(false);
-
-  const { message } = useSelector((state) => state.message);
-
-  const dispatch = useDispatch();
-
-  useEffect(() => {
-    dispatch(clearMessage());
-  }, [dispatch]);
-
-  const initialValues = {
-    username: userName,
-    password: "",
-  };
-
-  const validationSchema = Yup.object().shape({
-    username: Yup.string().required("This field is required!"),
-    password: Yup.string().required("This field is required!"),
-  });
-
-  const handleEditProfile = (formValue) => {
-    const { username, password } = formValue;
-
-    setLoading(true);
-
-    dispatch(editProfile({ username, password, accessToken }))
-      .unwrap()
-      .then(() => {
-        window.location.reload();
-      })
-      .catch(() => {
-        setLoading(false);
-      });
-  };
-
-  if (!accessToken) {
-    return <Navigate to="/login" />;
-  }
-
-  return (
-    <div className="container">
-      <header className="jumbotron">
-        <h1>Your Settings</h1>
-      </header>
-      <p>
-        <strong>email:</strong> {email}
-      </p>
-      <p>
-        <strong>username:</strong> {username}
-      </p>
-      <div className="col-md-12 login-form">
-        <div className="card card-container">
-          <img
-            src="//ssl.gstatic.com/accounts/ui/avatar_2x.png"
-            alt="profile-img"
-            className="profile-img-card"
-          />
-          <Formik
-            initialValues={initialValues}
-            validationSchema={validationSchema}
-            onSubmit={handleEditProfile}
-          >
-            {({ errors, touched }) => (
-              <Form>
-                <div className="form-group">
-                  <label htmlFor="email">Username</label>
-                  <Field
-                    name="username"
-                    type="text"
-                    className={
-                      "form-control" +
-                      (errors.username && touched.username ? " is-invalid" : "")
-                    }
-                  />
-                  <ErrorMessage
-                    name="username"
-                    component="div"
-                    className="invalid-feedback"
-                  />
-                </div>
-
-                <div className="form-group">
-                  <label htmlFor="password">Password</label>
-                  <Field
-                    name="password"
-                    type="password"
-                    className={
-                      "form-control" +
-                      (errors.password && touched.password ? " is-invalid" : "")
-                    }
-                  />
-                  <ErrorMessage
-                    name="password"
-                    component="div"
-                    className="invalid-feedback"
-                  />
-                </div>
-
-                <div className="form-group">
-                  <button
-                    type="submit"
-                    className="btn btn-primary btn-block"
-                    disabled={loading}
-                  >
-                    {loading && (
-                      <span className="spinner-border spinner-border-sm"></span>
-                    )}
-                    <span>Edit</span>
-                  </button>
-                </div>
-              </Form>
-            )}
-          </Formik>
-        </div>
-
-        {message && (
-          <div className="form-group">
-            <div className="alert alert-danger" role="alert">
-              {message}
-            </div>
-          </div>
-        )}
-      </div>
-    </div>
-  );
-};
-
-export default Settings;
