+++ src/app/pages/Settings/Settings.css
@@ -1,0 +1,25 @@
+.settings-container {
+  max-width: 1200px;
+  margin: 0 auto;
+}
+
+.settings-container .form-group .btn {
+  width: 183px;
+  height: 51px;
+  position: absolute;
+  margin-right: 14.5px;
+  right: 0;
+  background-color: #5cb85c;
+  border-color: #5cb85c;
+  font-size: 20px;
+}
+
+.settings-container .form-group textarea {
+  height: 226px;
+  min-height: 50px;
+}
+
+.invalid-feedback {
+  position: relative;
+  top: 3px;
+}
